refactor: context 코드 분리
